factory is now calling the PostProcessingCall
